# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ilm School Portal API
  version: 1.0.0
  description: |-
    Multi-school (multi-tenant) school management API. Every role except
    `system_admin` is scoped to exactly one school (`schoolId` on the JWT);
    `system_admin` is a platform-level role with no school of its own and
    can act across every school.

    **Auth**: session is an httpOnly `token` cookie (JWT), set by
    `POST /auth/login` and cleared by `POST /auth/logout`. There is no
    bearer-token flow — the cookie is sent automatically by the browser on
    same-origin requests (`withCredentials: true` on the frontend's axios
    client). A request with no cookie, an invalid/expired token, or a
    revoked token (see below) gets `401`.

    **Session lifetime**: tokens are signed for 1 hour. Any authenticated
    request past the halfway point of that window silently reissues the
    cookie, so an active user is never hard-logged-out mid-session — only
    genuine inactivity lets it expire. That sliding renewal is itself capped
    at 12 hours from the original login (`sessionStartedAt` in the token),
    so a session can't be kept alive indefinitely just by staying active —
    after 12 hours the user has to log in again regardless of activity.
    Logging out invalidates the token server-side immediately (via a
    `token_version` check on every request), not just client-side — and
    because there's no per-device session tracking, logging out on one
...
    sessions.

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /timetable/terms:
    get:
      tags:
        - Timetable
      summary: List school terms (admin/owner/parent view)
      description: Admin, owner, and parent. Same underlying `school_terms` table as `GET /report-cards/terms`, but its own role gate — report cards deliberately excludes owner, timetable deliberately includes it and also allows parents read access to term dates. Not gated by `timetable`, so the year view can be built before any slots exist.
      responses:
        '200':
          description: Terms
          content:
            application/json:
              schema:
                type: object
                properties:
                  terms:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        start_date:
                          type: string
                          format: date
                        end_date:
                          type: string
                          format: date
        '403':
          $ref: '#/components/responses/Forbidden'
      operationId: getTimetableTerms
    post:
      tags:
        - Timetable
      summary: Create a school term (admin/owner)
      description: Admin and owner only.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - start_date
                - end_date
              properties:
                name:
                  type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Term created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  term:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      start_date:
                        type: string
                        format: date
                      end_date:
                        type: string
                        format: date
        '400':
          description: Missing name/dates, or end_date before start_date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A term with this name already exists for the school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postTimetableTerms
  /timetable/slots:
    get:
      tags:
        - Timetable
      summary: List every class's timetable slots for one term (school-wide calendar view)
      description: Admin and owner only. Gated by `timetable`. Powers the weekly grid view — every class's slots at once for a given term, unlike `GET /timetable/classes/{classId}/slots` which is scoped to one class.
      parameters:
        - name: term_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        class_id:
                          type: integer
                        class_name:
                          type: string
                        term_id:
                          type: integer
                        slot_date:
                          type: string
                          format: date
                        start_time:
                          type: string
                        end_time:
                          type: string
                        subject_name:
                          type: string
                        teacher_id:
                          type: integer
                          nullable: true
                        teacher_first_name:
                          type: string
                        teacher_surname:
                          type: string
        '400':
          description: term_id missing, or doesn't belong to this school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
      operationId: getTimetableSlots
  '/timetable/classes/{classId}/slots':
    get:
      tags:
        - Timetable
      summary: List a class's timetable slots
      description: Admin and owner only. Gated by `timetable`. term_id is optional — omit it to see every slot for this class across every term.
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
        - name: term_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        term_id:
                          type: integer
                        term_name:
                          type: string
                        slot_date:
                          type: string
                          format: date
                        start_time:
                          type: string
                        end_time:
                          type: string
                        subject_name:
                          type: string
                          nullable: true
                        teacher_id:
                          type: integer
                          nullable: true
                        teacher_first_name:
                          type: string
                        teacher_surname:
                          type: string
        '400':
          description: term_id was provided but doesn't belong to this school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Not admin/owner, feature disabled, or class at a different school'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getTimetableClassesSlotsByClassId
    post:
      tags:
        - Timetable
      summary: Create a timetable slot for a class
      description: |-
        Admin and owner only. Gated by `timetable`. There's no term_id to
        set — it's resolved automatically from slot_date (whichever term's
        date range contains it). 400s if no term at this school covers
        that date. One slot per class/date/start_time.
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - slot_date
                - start_time
                - end_time
              properties:
                slot_date:
                  type: string
                  format: date
                start_time:
                  type: string
                  description: 'HH:MM'
                end_time:
                  type: string
                  description: 'HH:MM'
                subject_name:
                  type: string
                teacher_id:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Slot created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  slot:
                    type: object
                    properties:
                      id:
                        type: integer
                      term_id:
                        type: integer
        '400':
          description: Invalid date/time, no term at this school covers slot_date, or teacher_id isn't a teacher at this school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Not admin/owner, feature disabled, or class at a different school'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A slot already exists for this class at this date and start time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postTimetableClassesSlotsByClassId
  '/timetable/slots/{id}':
    put:
      tags:
        - Timetable
      summary: Edit a timetable slot
      description: Admin and owner only. Gated by `timetable`. No term_id — moving slot_date re-resolves which term the slot belongs to, the same way creation does.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - slot_date
                - start_time
                - end_time
              properties:
                slot_date:
                  type: string
                  format: date
                start_time:
                  type: string
                end_time:
                  type: string
                subject_name:
                  type: string
                teacher_id:
                  type: integer
                  nullable: true
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '400':
          description: Invalid date/time, no term at this school covers the new slot_date, or teacher_id isn't a teacher at this school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Not admin/owner, feature disabled, or slot''s class at a different school'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A slot already exists for this class at this date and start time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: putTimetableSlotsById
    delete:
      tags:
        - Timetable
      summary: Delete a timetable slot
      description: Admin and owner only. Gated by `timetable`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '403':
          description: 'Not admin/owner, feature disabled, or slot''s class at a different school'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: deleteTimetableSlotsById
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: httpOnly JWT set by POST /auth/login. See the top-level description for session lifetime/refresh/revocation behavior.
...
  responses:
    MessageOk:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
    Forbidden:
      description: 'Authenticated, but not permitted (wrong role, or outside the caller''s school)'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...